Cat6-Cat19: Security, best practices, architecture, and efficiency review
Code fixes applied:
- ota.cpp: Add #include "battery.h"; remove inline extern declarations from
  checkForFirmwareUpdate() (CAT7-01 — non-idiomatic pattern)
- ota.cpp: Replace goto read_headers with bool gotStatus flag in
  parseHttpResponseHeaders() (CAT7-02/CAT4-02 — goto elimination)
- main.cpp: Fix uint32_t overflow in DEBUG_NO_DEEP_SLEEP delay loop —
  use uint64_t for remainingMs (CAT12-01; debug-only path, no production impact)

Documented (no code change needed):
- CAT6-01: API key + OTA over HTTP — hardware limitation (SIM7000G TLS broken)
- CAT6-02: OTA proceeds without SHA-256 when .sha256 unavailable — intentional
  graceful degradation design choice
- CAT9-01: battery.h declares functions from main.cpp — known coupling, low risk

All Cat10-Cat19 checklist items verified correct and documented in review.

+++ src/main.cpp
@@ -798,11 +798,11 @@ void loop() {
 #if DEBUG_NO_DEEP_SLEEP
   SerialMon.println("DEBUG_NO_DEEP_SLEEP active: staying awake and delaying instead of deep sleep.");
   // Stay awake but idle for the sleep duration, resetting WDT periodically
-  uint32_t remainingMs = (uint32_t)sleepMinutes * 60UL * 1000UL;
+  uint64_t remainingMs = (uint64_t)sleepMinutes * 60ULL * 1000ULL; // uint64 avoids overflow at 129600 min
   const uint32_t chunkMs = 10000UL; // 10s chunks to keep logs responsive
   while (remainingMs > 0) {
     esp_task_wdt_reset();
-    uint32_t d = remainingMs > chunkMs ? chunkMs : remainingMs;
+    uint32_t d = (remainingMs > chunkMs) ? chunkMs : (uint32_t)remainingMs;
     delay(d);
     remainingMs -= d;
   }
